addDays and Month related type issue fixed

+++ app/Transaction.php
@@ -316,7 +316,7 @@ class Transaction extends Model
 //        ) {
 //            $transaction_date = \Carbon::parse($transaction->transaction_date);
 //            $number = (int) $transaction->pay_term_number;  // ✅ cast to int
-//            $due_date = $transaction->pay_term_type == 'days' ? $transaction_date->addDays($number) : $transaction_date->addMonths($number);
+//            $due_date = $transaction->pay_term_type == 'days' ? $transaction_date->addDays((int)$number) : $transaction_date->addMonths($number);
 //            $now = \Carbon::now();
 //            if ($now->gt($due_date)) {
 //                $payment_status = $payment_status == 'due' ? 'overdue' : 'partial-overdue';
@@ -333,9 +333,9 @@ class Transaction extends Model
     {
         $transaction_date = \Carbon::parse($this->transaction_date);
         if (! empty($this->pay_term_type) && ! empty($this->pay_term_number)) {
-            $due_date = $this->pay_term_type == 'days' ? $transaction_date->addDays($this->pay_term_number) : $transaction_date->addMonths($this->pay_term_number);
+            $due_date = $this->pay_term_type == 'days' ? $transaction_date->addDays((int)$this->pay_term_number) : $transaction_date->addMonths((int)$this->pay_term_number);
         } else {
-            $due_date = $transaction_date->addDays(0);
+            $due_date = $transaction_date->addDays((int)0);
         }
 
         return $due_date;
